project tree changed.
merge client_server and AccountManagement to the same project tree.

todo:
doxygen file does not change accordingly.

+++ src/common/error.c
@@ -60,7 +60,7 @@ void s_cserr_operation(void* data, enum cserr_op errop)
 
     if (errop == cserr_exit) {
         cssock_envclear();
-        exit(*(cserr_t*)data);
+        exit((int)*(cserr_t*)data);
     }
 
     errop &= (~cserr_exit);
